src/epam/final_task/Admitions-master/src/main/java/com/adminitions/admitions/auth/LogoutServlet.java
package com.adminitions.admitions.auth;

import jakarta.servlet.*;
import jakarta.servlet.http.*;
import jakarta.servlet.annotation.*;

import java.io.IOException;

@WebServlet(name = "/LogoutServlet", value = "/Logout")
public class LogoutServlet extends HttpServlet {
    @Override
    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
        HttpSession session = request.getSession();
        session.removeAttribute("User");
        session.removeAttribute("Name");
        response.sendRedirect("index.jsp");
    }
}
